Provide std from/into between color types.
